python/pyproject.toml
[project]
name = "leet1337"
version = "0.5.0"
description = "1337 inter-agent communication language — Python SDK"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
grpc = ["grpcio>=1.50", "grpcio-tools>=1.50"]
zmq = ["pyzmq>=25.0"]
websocket = ["websockets>=12.0"]
redis = ["aioredis>=2.0"]
yaml = ["pyyaml>=6.0"]
toml = ["tomli>=2.0", "tomli-w>=1.0"]
dev = [
    "pytest>=8",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "mypy",
    "ruff"
]
all = [
    "anthropic>=0.40",
    "openai>=1.0",
    "aiohttp>=3.8",
    "grpcio>=1.50",
    "grpcio-tools>=1.50",
    "pyzmq>=25.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "tomli>=2.0",
    "tomli-w>=1.0",
]

[project.scripts]
leet = "leet.cli:main"
leet-ide = "leet.adapters.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["leet*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
